requirements.txt
fastapi>=0.115.0
uvicorn[standard]==0.22.0
playwright>=1.44.0
playwright-recaptcha>=0.5.1
python-dotenv==1.0.0

